requirements.txt
langgraph
langchain
langchain-openai
langchain-google-genai
langsmith
fastapi
uvicorn
pandas
db-dtypes
python-dotenv
google-cloud-bigquery-storage
langchain-community
faiss-cpu
langchain-mistralai
rich
langchain-groq
google-cloud-bigquery[pandas]
aiofiles
tabulate
langchain-together
langchain_experimental
quickchart-io
boto3
awscli
langchain_aws
cloudinary
matplotlib